refactor: 메인페이지, 회원가입페이지 컴포넌트화
