+++ Frontend/src/pages/main.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Main() {
+  return (
+    <div>
+      main page
+    </div>
+  )
+}
+
+export default Main;
